#1262 implement path features
