+++ neuron_morphology/feature_extractor/data.py
@@ -1,3 +1,5 @@
+from typing import Union
+
 from neuron_morphology.morphology import Morphology
 
 class Data:
@@ -11,3 +13,11 @@ class Data:
         self.morphology: Morphology = morphology
         for name, value in other_things.items():
             setattr(self, name, value)
+
+
+MorphologyLike = Union[Data, Morphology]
+
+def get_morphology(data: MorphologyLike):
+    if isinstance(data, Morphology):
+        return data
+    return data.morphology
